+++ R/zzz.R
@@ -5,22 +5,39 @@
   invisible()
 }
 
-#' @importFrom usethis ui_todo ui_code ui_oops ui_yeah ui_done ui_stop
+#' @importFrom usethis ui_todo ui_code ui_code_block ui_oops ui_yeah ui_done
+#'   ui_stop
 #'
 #' @description \code{check_pat} can be used to check if you have a GITHUB_PAT
 #'   set and will advise on how to do this if necessary.
+#' @param create Whether to help you create a personal GITHUB_PAT if you do not
+#'   have one set. When \code{create=FALSE} a default PAT will be used if you
+#'   have not set your own. This could cause trouble if other people are using
+#'   the same PAT.
 #' @export
+#' @return \code{check_pat} returns the PAT invisibly or errors out if
+#'   \code{create=TRUE} and none can be set.
 #' @rdname install
 #' @examples
 #' \dontrun{
-#' # Check status of GitHub PAT
-#' check_pat()
+#' # Check status of GitHub PAT and create one if required
+#' natmanager::check_pat(create=FALSE)
+#' # Check status of GitHub PAT and use default if no personal one available
+#' natmanager::check_pat(create=FALSE)
 #' }
-check_pat <- function() {
+check_pat <- function(create=TRUE) {
   pat=usethis::github_token()
-  if (isFALSE(nzchar(pat))) {
+  if (isFALSE(nzchar(pat))  && isFALSE(create)) {
+    pat <- paste0("aa01baa46278a",
+                  "ab50961f28153",
+                  "26e77befcf94c3")
+    ui_todo(c(
+      "Using natmanager's default GitHub access token (PAT) to install packages.",
+      "If you see error messages relating to GitHub rate limiting then use:",
+      "{ui_code('natmanager::check_pat()')}","to create your own."))
+  } else if (isFALSE(nzchar(pat))) {
     if (interactive()) {
-      ui_oops("You must have a GitHub account and PAT (Personal Access Token) to install the full natverse!")
+      ui_oops("You should have a GitHub account and PAT (Personal Access Token) to install the full natverse!")
       ui_todo("Please read http://natverse.org/install/ for details!")
       message("")
       res=ui_yeah("Shall I open the help page for you?")
@@ -42,8 +59,9 @@ check_pat <- function() {
            "See http://natverse.org/install/ for details!")
     }
   } else{
-    # nothing to do right now
-    ui_done("GITHUB_PAT is set.")
-    invisible(TRUE)
+    # nothing to do. Only put message if we were considering making a PAT
+    if(create)
+      ui_done("You have a personal GITHUB_PAT set.")
   }
+  invisible(pat)
 }