+++ NEWS.md
@@ -1,3 +1,5 @@
+# natmanager (development version)
+
 # natmanager 0.3.0
 
 * option to